% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_summed_exps.R
\name{log_summed_exps}
\alias{log_summed_exps}
\title{This is a computational function}
\usage{
log_summed_exps(x)
}
\arguments{
\item{x}{A vector of numeric values}
}
\value{
Returns the logarithm of the sum of exponents of the given vector.
All NA values are automatically removed from computation.
}
\description{
This is a computational function
}
\examples{
data = c(1:200)
log_summed_exps(x = data)

}
\author{
Awan Afiaz
}
